refactor to render alerts from container
