+++ src/applications/medical-copays/components/AlertPage.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Alert from '../components/Alerts';
+
+const AlertPage = ({ pathname, alertType, error }) => {
+  const overview = 'Your current copay balances';
+  const details = 'Copay bill details';
+
+  return (
+    <>
+      <h1>{pathname === '/' ? overview : details}</h1>
+      <Alert type={alertType} error={error} />
+    </>
+  );
+};
+
+export default AlertPage;
